Update TestApplicationPage, Application legacy
