+++ packages/hello-gsm/src/PageContainer/ApplicationPage/style.ts
@@ -100,19 +100,14 @@ export const Table = styled.table`
     width: 18.7vh;
     height: 25vh;
   }
+`;
 
-  svg {
-    width: 100%;
-    height: 100%;
-  }
+export const Slash = styled.td`
+  background: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg"><line x1="0" y1="100%" x2="100%" y2="0" stroke="gray" /></svg>');
+`;
 
-  .slash {
-    background: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg"><line x1="0" y1="100%" x2="100%" y2="0" stroke="gray" /></svg>');
-  }
-
-  .backSlash {
-    background: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg"><line x1="0" y1="0" x2="100%" y2="100%" stroke="gray" /></svg>');
-  }
+export const BackSlash = styled.td`
+  background: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg"><line x1="0" y1="0" x2="100%" y2="100%" stroke="gray" /></svg>');
 `;
 
 export const Subject = styled.td`
@@ -184,8 +179,6 @@ export const Materials = styled.div`
   line-height: 3.5vh;
 `;
 
-export const ScoreDetails = styled.div``;
-
 export const SignatureSection = styled.div`
   display: flex;
   justify-content: space-between;
@@ -197,13 +190,6 @@ export const ScoreTable = styled.div`
   display: flex;
   height: 26vh;
   border: 0.08vh solid #000000;
-  .slash {
-    background: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg"><line x1="0" y1="100%" x2="100%" y2="0" stroke="gray" /></svg>');
-  }
-
-  .backSlash {
-    background: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg"><line x1="0" y1="0" x2="100%" y2="100%" stroke="gray" /></svg>');
-  }
 `;
 
 export const GradeAndSubject = styled.div`
@@ -249,16 +235,15 @@ export const DivSubject = styled.div`
   height: 2.2vh;
 `;
 
-export const Slash = styled.div`
+export const DivSlash = styled.div`
   height: 100%;
   background: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg"><line x1="0" y1="0" x2="100%" y2="100%" stroke="gray" /></svg>');
   border-bottom: 0.08vh solid #000000;
 `;
 
-export const TestTable = styled.div`
+export const NonScoreTable = styled.div`
   display: flex;
   flex-direction: column;
-  height: 26vh;
 `;
 
 export const ColumnWrapper = styled.div`
@@ -269,11 +254,53 @@ export const ColumnWrapper = styled.div`
   width: 100%;
 `;
 
-export const Test = styled.div`
+export const ConversionPoint = styled.div`
   display: flex;
   text-align: center;
   border: 0.08vh solid #000000;
   border-top: 0;
   height: 2.6vh;
   line-height: 2.6vh;
+`;
+
+export const PrintBtn = styled.button`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 1.8vh 2.5vh;
+  position: fixed;
+  width: 17vh;
+  height: 6.5vh;
+  right: 9vh;
+  bottom: 9vh;
+  border-radius: 0.7vh;
+  box-shadow: 0.3vh 0.3vh 0.5vh rgba(0, 0, 0, 0.21);
+  z-index: 7;
+  box-sizing: border-box;
+  cursor: pointer;
+  background-color: #ffffff;
+  border: none;
+
+  svg {
+    width: 2.5vh;
+  }
+  &:hover {
+    transition: 0.5s;
+    background-color: #000000;
+    p {
+      color: #ffffff;
+    }
+    svg > path {
+      fill: #ffffff;
+    }
+  }
+
+  @media print {
+    display: none;
+  }
+`;
+
+export const PrintDesc = styled.p`
+  font-size: 2.1vh;
+  font-weight: bold;
 `;
