+++ packages/hello-gsm/src/PageContainer/index.tsx
@@ -13,4 +13,3 @@ export { default as CalculatorPage } from './CalculatorPage';
 export { default as TestCalculatorPage } from './TestPage';
 export { default as InspectionPage } from './InspectionPage';
 export { default as GEDPage } from './GEDPage';
-export { default as TestApplicationPage } from './TestApplicationPage';
